+++ R/toolbar.R
@@ -1,0 +1,28 @@
+#' Toolbar component
+#'
+#' @description
+#' A toolbar which can contain buttons, inputs, and other UI elements in a small
+#' form suitable for inclusion in card headers, footers, and other small places.
+#'
+#' @param ... UI elements for the toolbar.
+#' @param align Determines if toolbar should be aligned to the `"right"` or
+#'   `"left"`.
+#' @return Returns a toolbar element.
+#'
+#' @export
+toolbar <- function(
+  ...,
+  align = c("right", "left")
+) {
+  align <- rlang::arg_match(align)
+
+  tag <- div(
+    class = "bslib-toolbar bslib-gap-spacing",
+    "data-align" = align,
+    ...,
+    component_dependencies()
+  )
+
+  tag_require(tag, version = 5, caller = "toolbar()")
+  as_fragment(tag)
+}